+++ CHANGELOG.md
@@ -7,6 +7,15 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 
 ## [Unreleased]
 
+### Added
+
+- `deploy add status` command
+- `deploy trigger` command
+
+### Changed
+
+- deprecate `deploy` command
+
 ## [v0.14.0] - 2024-07-25
 
 ### Added